change function type

+++ src/3pipe.h
@@ -11,6 +11,6 @@
 bool isPrime(const unsigned long);
 void send(const unsigned int);
 unsigned int transform(unsigned int);
-long int recive();
+unsigned int recive();
 
 #endif //__3PIPE_H_
